+++ alertClientExecuter/lib/alert/executer.py
@@ -8,7 +8,6 @@
 # Licensed under the GNU Affero General Public License, version 3.
 
 import os
-import logging
 import subprocess
 import json
 from typing import List
@@ -24,7 +23,7 @@ class ExecuterAlert(_Alert):
         _Alert.__init__(self)
 
         # used for logging
-        self.fileName = os.path.basename(__file__)
+        self._log_tag = os.path.basename(__file__)
 
         # Flag that indicates if command execution for corresponding received message is activated.
         self.cmd_triggered_activated = False
@@ -60,14 +59,13 @@ class ExecuterAlert(_Alert):
 
     def _execute_cmd(self, execute_cmd_list: List[str]):
 
-        logging.debug("[%s]: Alert '%d' executing command '%s'." % (self.fileName, self.id, " ".join(execute_cmd_list)))
+        self._log_debug(self._log_tag, "Executing command '%s'." % " ".join(execute_cmd_list))
 
         try:
             subprocess.Popen(execute_cmd_list, close_fds=True)
         except Exception as e:
-            logging.exception("[%s]: Alert '%d' executing process failed."
-                              % (self.fileName, self.id))
-            logging.error("[%s]: Alert '%d' command was: %s" % (self.fileName, self.id, " ".join(execute_cmd_list)))
+            self._log_exception(self._log_tag, "Executing process failed.")
+            self._log_error(self._log_tag, "Command was: %s" % " ".join(execute_cmd_list))
 
     def initialize(self):
         """
@@ -99,7 +97,7 @@ class ExecuterAlert(_Alert):
         if not self.cmd_triggered_activated:
             return
 
-        logging.info("[%s]: Alert '%d' executes for state '%d'." % (self.fileName, self.id, sensor_alert.state))
+        self._log_info(self._log_tag, "Executing command for state '%d'." % sensor_alert.state)
 
         # Prepare command execution by replacing all placeholders.
         temp_execute = list(self.cmd_triggered_list)
@@ -121,7 +119,7 @@ class ExecuterAlert(_Alert):
         if not self.cmd_normal_activated:
             return
 
-        logging.info("[%s]: Alert '%d' executes for state '%d'." % (self.fileName, self.id, sensor_alert.state))
+        self._log_info(self._log_tag, "Executing command for state '%d'." % sensor_alert.state)
 
         # Prepare command execution by replacing all placeholders.
         temp_execute = list(self.cmd_normal_list)
@@ -147,7 +145,7 @@ class ExecuterAlert(_Alert):
         if profile.profileId not in self.cmd_profile_change_target_profiles:
             return
 
-        logging.info("[%s]: Alert '%d' executes for profile change message." % (self.fileName, self.id))
+        self._log_info(self._log_tag, "Executing for profile change.")
 
         # Prepare command execution by replacing all placeholders.
         temp_execute = list(self.cmd_profile_change_list)