+++ alertClientKodi/lib/alert/kodi.py
@@ -9,7 +9,6 @@
 
 import time
 import os
-import logging
 import kodijson
 from .core import _Alert
 from ..globalData import ManagerObjSensorAlert, ManagerObjProfile
@@ -22,7 +21,7 @@ class KodiAlert(_Alert):
     def __init__(self):
         _Alert.__init__(self)
 
-        self.fileName = os.path.basename(__file__)
+        self._log_tag = os.path.basename(__file__)
 
         # these values are used to check when the alert was triggered
         # the last time and if it should trigger again
@@ -57,8 +56,7 @@ class KodiAlert(_Alert):
             self.triggered = utc_timestamp
 
             intersect_alert_levels = [str(x) for x in set(sensor_alert.alertLevels).intersection(self.alertLevels)]
-            logging.info("[%s] Alert '%d' triggered for alert levels %s."
-                         % (self.fileName, self.id, ", ".join(intersect_alert_levels)))
+            self._log_info(self._log_tag, "Triggered for alert levels %s." % ", ".join(intersect_alert_levels))
 
             # extract the received message if it was received and should be
             # displayed
@@ -73,7 +71,7 @@ class KodiAlert(_Alert):
                 kodi_obj = kodijson.Kodi("http://" + self.host + ":" + str(self.port) + "/jsonrpc")
 
             except Exception as e:
-                logging.exception("[%s]: Alert '%d' not able to connect to Kodi instance." % (self.fileName, self.id))
+                self._log_exception(self._log_tag, "Not able to connect to Kodi instance.")
                 return
 
             # ping the kodi instance
@@ -81,7 +79,7 @@ class KodiAlert(_Alert):
                 response = kodi_obj.JSONRPC.Ping()
 
             except Exception as e:
-                logging.exception("[%s]: Alert '%d' not able to ping Kodi instance." % (self.fileName, self.id))
+                self._log_exception(self._log_tag, "Not able to ping Kodi instance.")
                 return
 
             # check if kodi instance respond
@@ -135,7 +133,7 @@ class KodiAlert(_Alert):
                                                   displaytime=self.displayTime)
 
             else:
-                logging.error("[%s]: Alert '%d' Kodi does not respond." % (self.fileName, self.id))
+                self._log_error(self._log_tag, "Kodi does not respond.")
 
     def initialize(self):
         """